+++ src/main/java/br/edu/tglima/locadora/controllers/CadCliente.java
@@ -8,6 +8,7 @@ import javax.inject.Named;
 
 import br.edu.tglima.locadora.models.pessoa.Cliente;
 import br.edu.tglima.locadora.models.pessoa.OpGeneros;
+import br.edu.tglima.locadora.util.ClienteUtil;
 import br.edu.tglima.locadora.util.FacesUtil;
 
 @Named
@@ -22,24 +23,11 @@ public class CadCliente implements Serializable {
 	public void cadastrar(){
 		
 		FacesUtil.enviarMsgSucesso(null, "Cliente cadastrado com sucesso!");
-		exibirDadosInformados();
+		System.out.println("::::::::::::::::::::::::::::::::::::: CLIENTE SALVO :::::::::::::::::::::::::::::::::::::");
+		ClienteUtil.exibirDadosNoConsole(nc);
 		this.nc = new Cliente();
 	}
 	
-//	TODO Eliminar este método depois
-	private void exibirDadosInformados(){
-		System.out.println("###########################################################");
-		System.out.println("                   Dados informados");
-		System.out.println("Nome do cliente: " + this.nc.getNome() );
-		System.out.println("Sobrenome: " + this.nc.getSobrenome() );
-		System.out.println("Gênero: " + this.nc.getGenero() );
-		System.out.println("Data de nascimento: " + this.nc.getDataNasc() );
-		System.out.println("Telefone: " + this.nc.getTelefone());
-		System.out.println("Nº Registro CNH: " + this.nc.getHabilitacao());
-		System.out.println("Data de validade da CNH: " + this.nc.getValidadeHab());
-		System.out.println("###########################################################\n");
-	}
-
 	//Getters de acesso aos Enums
 	public OpGeneros[] getGeneros(){
 		return OpGeneros.values();